feat(part4): change blog delete for specific user
only a user with a valid token and who has created the blog will be able to delete it

+++ part4/bloglist/controllers/blogs.js
@@ -43,11 +43,27 @@ blogsRouter.get('/:id', async (request, response) => {
   const id = request.params.id
   const blog = await Blog.findById(id)
 
+  if (!blog) {
+    return response.status(404).end()
+  }
+
   return response.status(200).json(blog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(id)
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'user not authorized to delete this blog' })
+  }
+
   await Blog.findByIdAndRemove(id)
 
   return response.status(204).end()
